refactor API and EXT trait a bit
