+++ crates/iota-core/src/generate_format.rs
@@ -22,7 +22,7 @@ use iota_types::{
     },
     digests::ConsensusCommitDigest,
     effects::{
-        IDOperation, ObjectIn, ObjectOut, TransactionEffects, TransactionEffectsAPIForTesting,
+        IDOperation, ObjectIn, ObjectOut, TransactionEffects, TransactionEffectsExt,
         TransactionEvents, UnchangedSharedKind,
     },
     event::Event,
@@ -560,7 +560,7 @@ fn get_registry() -> Result<Registry> {
     // Trace FullCheckpointContents, CheckpointTransaction and CheckpointData
     // via trace_value (they transitively contain TypeTag).
     let sample_transaction = Transaction::new(sender_data.clone());
-    let sample_effects = TransactionEffects::empty_for_testing(TransactionDigest::default());
+    let sample_effects = TransactionEffects::new_empty_v1(TransactionDigest::default());
     let sample_exec_data = ExecutionData {
         transaction: sample_transaction.clone(),
         effects: sample_effects.clone(),
