+++ crates/iota-data-ingestion-core/src/tests.rs
@@ -19,7 +19,7 @@ use iota_types::{
     committee::EpochId,
     crypto::KeypairTraits,
     digests::ObjectDigest,
-    effects::{TransactionEffects, TransactionEffectsAPIForTesting},
+    effects::{TransactionEffects, TransactionEffectsExt},
     full_checkpoint_content::{CheckpointData, CheckpointTransaction},
     gas::GasCostSummary,
     messages_checkpoint::{
@@ -451,7 +451,7 @@ async fn basic_flow_with_custom_callback() {
     );
 
     let transaction = Transaction::from_data(tx_data, vec![]);
-    let effects = TransactionEffects::empty_for_testing(*transaction.digest());
+    let effects = TransactionEffects::new_empty_v1(*transaction.digest());
     let ch_tx = CheckpointTransaction {
         transaction,
         effects,
