+++ crates/iota-json-rpc/src/coin_api.rs
@@ -578,7 +578,7 @@ mod tests {
         base_types::{IotaAddress, ObjectID, SequenceNumber, StructTag, TypeTag},
         coin::TreasuryCap,
         digests::{ObjectDigest, TransactionDigest},
-        effects::{TransactionEffects, TransactionEffectsAPIForTesting, TransactionEvents},
+        effects::{TransactionEffects, TransactionEffectsExt, TransactionEvents},
         error::{IotaError, IotaResult},
         id::UID,
         messages_checkpoint::{CheckpointDigest, CheckpointSequenceNumber},
@@ -1379,7 +1379,7 @@ mod tests {
         #[tokio::test]
         async fn test_object_not_found() {
             let transaction_digest = TransactionDigest::from([0; 32]);
-            let transaction_effects = TransactionEffects::empty_for_testing(transaction_digest);
+            let transaction_effects = TransactionEffects::new_empty_v1(transaction_digest);
 
             let mut mock_state = MockStateRead::new();
             mock_state
@@ -1506,7 +1506,7 @@ mod tests {
             let package_id = get_test_package_id();
             let (coin_name, _, _, _, _) = get_test_treasury_cap_peripherals(package_id);
             let transaction_digest = TransactionDigest::from([0; 32]);
-            let transaction_effects = TransactionEffects::empty_for_testing(transaction_digest);
+            let transaction_effects = TransactionEffects::new_empty_v1(transaction_digest);
 
             let mut mock_state = MockStateRead::new();
             mock_state
